Ensure addVariant actions are called at the end.
Example where this is necessary is: You have product with a SameForAll
constrained attribute, which you want to change and at the same time
add another variant to this products.

In order to import both changes successfully in one POST, we have to first update the
attributes value and then create the variant. Otherwise the variant would be added
with the new value for the attribute, but the backend would validate that it does
not fit to the existing value of all existing variants.